# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/cluster/pgsqlcluster": {
            "post": {
                "description": "CreateCluster deploy postgres cluster to the kubernetes cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "CreateCluster deploy postgres cluster to the kubernetes cluster.",
                "parameters": [
                    {
                        "description": "Deploy pgsqloperator params",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pgsql.CreateClusterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"OK\",\"data\":{}}",
                        "schema": {
                            "$ref": "#/definitions/pgsql.CreateClusterResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "DeleteCluster delete postgres clsuter from the kubernetes cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "DeleteCluster delete postgres cluster from the kubernetes cluster.",
                "parameters": [
                    {
                        "description": "Delete postgres cluster params",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pgsql.DeleteClusterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"OK\",\"data\":{}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            }
        },
        "/cluster/pgsqlcluster/detail/{name}/{namespace}": {
            "get": {
                "description": "GetCluster get a postgres cluster information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "GetCluster get a postgres cluster information.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "postgres cluster name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"message\":\"OK\",\"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            }
        },
        "/cluster/pgsqlcluster/list/{namespace}": {
            "get": {
                "description": "GetClusterList get the list of the postgres cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "GetClusterList get the list of the postgres cluster.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"message\":\"OK\",\"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            }
        },
...
        "/operator/mysqloperator": {
            "post": {
                "description": "安装 mysql operator.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operator"
                ],
                "summary": "安装 mysql operator.",
                "parameters": [
                    {
                        "description": "安装MySQL Operator组件时所需参数.",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.CreateOperatorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"OK\",\"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "从Kubernetes集群中删除mysql operator组件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operator"
                ],
                "summary": "从Kubernetes集群中删除mysql operator组件",
                "parameters": [
                    {
                        "description": "删除MySQL Operator组件时所需的参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.DeleteOperatorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"message\":\"OK\",\"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mysql.CreateOperatorRequest": {
            "type": "object",
            "properties": {
                "namespace": {
                    "description": "Namespace Mysql Operator 安装在哪个命名空间之下.",
                    "type": "string"
                }
            }
        },
...
        "mysql.DeleteOperatorRequest": {
            "type": "object",
            "properties": {
                "namespace": {
                    "description": "Namespace Mysql Operator 组件安装在哪个命名空间之下.",
                    "type": "string"
                }
            }
        },
...
        "pgsql.CreateClusterRequest": {
            "type": "object",
            "properties": {
                "ClusterName": {
                    "type": "string"
                },
                "Databases": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "Namespace": {
                    "type": "string"
                },
                "Postgresql": {
                    "type": "object",
                    "$ref": "#/definitions/pgsql.PostgresqlParam"
                },
                "Replicas": {
                    "type": "integer"
                },
                "TeamId": {
                    "type": "string"
                },
                "Users": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/pgsql.UserFlags"
                    }
                },
                "Volume": {
                    "type": "object",
                    "$ref": "#/definitions/pgsql.Volume"
                }
            }
        },
        "pgsql.CreateClusterResponse": {
            "type": "object",
            "properties": {
                "ClusterName": {
                    "type": "string"
                },
                "Databases": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "Namespace": {
                    "type": "string"
                },
                "Postgresql": {
                    "type": "object",
                    "$ref": "#/definitions/pgsql.PostgresqlParam"
                },
                "Replicas": {
                    "type": "integer"
                },
                "TeamId": {
                    "type": "string"
                },
                "Users": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/pgsql.UserFlags"
                    }
                },
                "Volume": {
                    "type": "object",
                    "$ref": "#/definitions/pgsql.Volume"
                }
            }
        },
...
        "pgsql.DeleteClusterRequest": {
            "type": "object",
            "properties": {
                "ClusterName": {
                    "type": "string"
                },
                "Namespace": {
                    "type": "string"
                }
            }
        },
        "pgsql.PostgresqlParam": {
            "type": "object",
            "properties": {
                "parameters": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "pgsql.UserFlags": {
            "type": "array",
            "items": {
                "type": "string"
            }
        },
        "pgsql.Volume": {
            "type": "object",
            "properties": {
                "Size": {
                    "type": "string"
                },
                "StorageClass": {
                    "type": "string"
                },
                "SubPath": {
                    "type": "string"
                }
            }
        },
...
        "tool.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}